Moved some planning documents to a seperate folder
Even though we're using the google drive, it sounded like the class requires
that we have the files on the github as well.

Currently, the files on the git hub are a little out of date (compared to the
google drive), so I'm going to update them later.
update them later.