Reworked equipment system
Equipment loadouts are now stored as lists of item ids. The index in the
equipment list decides what kind of item it is. For example, if my equip list
if [0, 7, 4], then I have the item with id 7 as my weapon (which is index 1),
and the item with id 4 as my off-hand (which is index 2). This index-type
mapping is stored in the database table 'equip types'.

The 0th index is unused. Also, if an index is 0 (other than the 0th index),
then there is nothing equipped to that slot.