Furhter research on communication between Unity and web2py
Had to dig through the Unity Networking APIs for a bit, but I found a way for
the unity project to send a webrequest to the web server. Now I just have to
implement a way for the server to know who is asking for the request so that
I can match up the stats properly.